♻️ Fix all broken tests caused by the migration of the customized checking point abstraction.
